Fix various bugs

+++ include/Buffer.hpp
@@ -9,20 +9,20 @@ namespace md_audio {
 
     class Buffer {
     public:
-        explicit Buffer(memory::Allocatable<MdFloat*>&, std::uint32_t);
+        explicit Buffer(memory::Allocatable<MdFloat*>&, std::size_t);
 
         void initialise();
 
-        MdFloat& operator[] (std::uint32_t) noexcept;
+        MdFloat& operator[] (std::size_t) noexcept;
 
-        const MdFloat& operator[] (std::uint32_t) const noexcept;
+        const MdFloat& operator[] (std::size_t) const noexcept;
 
         ~Buffer();
 
     private:
         memory::Allocatable<MdFloat*>& m_allocator;
         MdFloat* m_memory = nullptr;
-        const std::uint32_t m_size;
+        const std::size_t m_size;
     };
 
 }
